`lsp-rename': highlight renaming
`lsp-rename' now highlights the identifier being renamed.

To do so, refactor `lsp--get-symbol-to-rename' to return the bounds of
the symbol being renamed instead of it, and allow its placeholder to be
nil. The latter is necessary to reduce code duplication, as the two code
paths of that function (has prepareRename, doesn't have prepareRename)
would need to do the same `buffer-substring' with the same bounds.

Add a new function, `lsp--read-rename', which takes the result of a call
to `lsp--get-symbol-to-rename' and either `user-error's (if the result
was nil) or asks the user for a new symbol, highlighting the identifier
being renamed. The split was done to make them independently testable.

`lsp-rename': no longer `user-error' if NEW-NAME is nil, since that only
makes sense if it is called interactively. However, in the latter case,
`lsp--read-rename' would do the necessary error handling already.