Use a consistent placeholder face
`lsp-rename-placeholder-face' is now used to display the new name
instead of reusing the face of the identifier in the buffer.

The reasoning here is that reusing the symbols's face directly would
yield inconsistent placeholders, while conveying no additional
information, given that the thing being renamed is shown in that very
same prompt as it appears in the buffer.